pyproject.toml
[project]
name = "telegram-scraper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.1",
    "google-cloud-logging==3.12",
    "google-cloud-secret-manager==2.23.3",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "telethon==1.39.0",
    "uvicorn==0.34.2",
]
